Add concern to define phases in models
The phases are defined with a set of columns of the model table.
The concern creates two types of phases: regular phases with start date,
end date and an attribute to enable or disable the phase and
publications, which are phases with no end date. The concern provides
methods to validate dates, get the status of the phase and some scopes
with the name of the phase.