20170921 - Refactored residence verification
The goal for this issue was to create a setting that allow forks to easily disable verification againt geographical zones.

For that matter, a setting was created on seeds file, which set to true allows to perform residential verification via census.

On the other hand, if set to false, overrides the verification and goes trough SMS verification.

 On branch mlucena-refactor-residence-verification
 Changes to be committed:
	modified:   app/controllers/application_controller.rb
	modified:   app/controllers/verification/residence_controller.rb
	modified:   app/controllers/verification_controller.rb
	modified:   config/locales/en/verification.yml
	modified:   config/locales/es/verification.yml
	modified:   db/dev_seeds.rb
	modified:   db/seeds.rb