Redirect to budget's index when no filter params
We still need to figure out how to nicely navigate between feasible,
unfeasible and selected proposals

Currently the budget’s show page displays a budget’s groups but in a
very minimal way. The budget’s groups are also displayed in the
budget’s index page in a much nicer way

Thus, we want users to navigate from the nicer budget’s index page if
possible. The only case that it is not possible is if they want to
filter by unfeasible or selected projects. Therefore, if the url has
filter params we display the budget’s show page, otherwise we redirect
them to the budget’s index page