Try if notifiable has notifiable methods
Why:

There are Notifications with associated `notifiables` that actually are
not anymore Notifiables (the class doesn't include the Notifiable
concern). So when Notification delegates certain "notifiable" methods
to them the is an error.

How:

Using `try` directly on the notifiable association to avoid the delegate
trap on those corner case scenarios.