Add the UI for the feature
A new button has been added to the
management account page. When it is
clicked, the reset password options are shown.
If the user has an email, a button to send
a reset password email appears. If there is no email, a form
to reset the password by hand appears.

When the password has been changed by hand,
it can be printed using a button in the screen.