Fix assigned investments count in valuation panel
Why:

At the valuation panel the number of "Investment projects
assigned with valuation open" for each budget was only taking into
account the individually assigned investments to the logged in valuator.

That count should include as well investments assigned to the valuator
group the valuator belongs to (if any)

How:

* Using the new Budget::Investment#accesible_by_valuator scope as well
as checking if the current user is a valuator
* Improving the test coverage for the valuation budget view to cover the
count number.