Upstream (#38)
* Improve newsletter model spec with valid_segment_recipient?

* Improve newsletter model spec list_of_recipients

* Disable guide page when Budget is not accepting

Why:

When there is not Budget accepting (Investment creation) the guide page
doesn't have much sense as it will give the user an option that can't be
used (creating an Investment).

How:

Using `Budget.current&.accepting?` conditional at GuidesHelper to link
to new proposal link instead of guide page, and adding an scenario to
guides feature spec for it.

* docs(readme): Include Node.js as requirement

Failed CONSUL installation reported on #2391 because
Node.js was missing on the development environment

* Improved user segments en & es translations

* Fixes button vertical align on polls list

* Adds missing translation for legislation filter title

* Refactors documents partials and reorganices css

* Removes unnecessary css

* Removes document tab on proposals

* Removes document tab on budgets investments

* Removes duplicate heading info, fixes headings structure

* Removes document tab on legislation proposals

* Updates specs

* Fixes documentable rubocop offences

* Added unfeasible investments with valuation not finished to `UserSegments#feasible_and_undecided_investment_authors`

* Changed tests for `UserSegments#feasible_and_undecided_investment_authors`

* Changed translations

* Changed how newsletters controller and mailer handle recipients

Now newsletters controller calls the Mailer method to send a newsletter once per user.

* Fixed newsletter tests

* Update unreleased section of changelog

* Improve download segment translations

* Remove empty emails from users segments recipients

Why:

User with an empty email value (nil) should not appear in the recipient
list for a given UserSegment at Newsletters or Email Downloads.

How:

Using Enumerable#compact and Enumerable#select to filter out empty emails

Increasing Email Download feature spec and Newsletter model spec to cover
all possible scenarios including the nil email one.

* Add UserSegments#user_segment_emails helper method

Why:

Both Newsletters and Email Downloads need the same logic: To extract the
emails from all the users in the segment that have newsletter flag
active, removing all empty email values.

How:

1- UserSegments#user_segment_emails holds that repeated logic and is used
on both Newsletter & EmailDownload.

2- Rename Newsletter#list_of_recipients to list_of_recipient_emails as
it is more descriptive. There is no need to pass entire Users around,
only the emails are needed at Mailer#newsletter method.

3- Cleanup Newsletter#list_of_recipient_emails model spec scenario

* The function comment_on has been modified to change the functionality. Now, a comment is generated and a email notification send manually, so that JS has been avoided, that was the original problem.

* first changed for click_link so that the test can wait the capybara default time.

* Make config.time_zone configurable at secrets.yml

* Improve Github's Pull Request Template file

The actual PR template feels more like bureaucracy than an actual guide
or checklist to help the PR author explain all important thigs that any
reviewer or changelog reader may need to understand.

We'll be moving most of the redundant things (like remembering tests are
needed, or explaning how things where implemented with a clear and
granular commit history) into a Wiki/Doc entry.

For regular contributors there is no need for reminders, we need to
improve new contributors landing with good guides and lowering the bar
for first PR's

* Swap @valuator instance variable for a let

Why:

Avoid instance variables as we agreed upon with RSpec/InstanceVariable

How:

Using a let(:valuator) and replacing all `@valuator` with `valuator`

* Swap @budget instance variable for a let

Why:

Avoid instance variables as we agreed upon with RSpec/InstanceVariable

How:

Using a let(:budget) and replacing all `@budget` with `budget`

* Swap @investment instance variable for a let

Why:

Avoid instance variables as we agreed upon with RSpec/InstanceVariable

How:

Using a let(:invsetment) and replacing all `@investment` with
`investment`, as well as adding a let(:admin) for the administrator.

* Refactor Valuator spec Show context with lets

Extract duplicated code used to set the scenario conditions to simple
lets (admin, second valuator and investment)

* Make valuator spec 100 LineLength compliant

Readability of the code is important

* Trim valuator investment valuation ability scope

Valuators shouldn't be able to valuate an investment which valuation has
already been finished (only admins can)

* Prevent valuators from editing finished valuation

Valuators should not be able to edit a finished valuation (only admins
should).

The valuation form is only shown to the valuator if he has that ability
(we've previously modified app/models/abilities/valuator.rb to be able
to rely on `valuate` over an investment to check that)

If the valuator can't see the form, we present him just the data in
plain text.

* Adds correct url on budget investment twitter share

* Adds share buttons and refactors processes header

* Removes unused js

* Removes unused css

* Uses rem-calc function instead of explicit units

* Removes sandbox folder

* Removes unused css

* Adds specs for additional info button

* Adds description to legislation process share buttons

* Replaces hardcoded social share buttons tag with correct partial

* Adds description for social share buttons

* Use user locale instead of default locale to format currencies

* Clean css and views of custom content

* Adds legislation processes styles

* Add Budget#formatted_amount unit test

* Adds heading link on different heading message

* Extract BudgetHeadingsHelper#heading_link & use it

Why:

The logic to construct the link to a heading (if it exists) is in three
different places, this is a clear candidate for a helper method.

How:

Just checking at the helper method if `assigned_heading` and `budget`
has values and composing the link if so.

* Makes smaller text on participation not allowed div

* Replace format_price for Budget#formatted_amount

ApplicationHelper#format_price and Budget#formatted_amount has the same
objective and code, but the Budget#formatted_amount method also uses the
currency of the Budget to correctly give currencies format.

By replacing usage of format_price with formatted_amount we can remove
format_price and have a single location for currency format logic.

* Avoid explicity ID values on fixtures if possible

The specs don't care about the actual ID value, only about using the
same value. Fixed ID values cause flaky specs and high maintenance issues

* Update CHANGELOG v0.14

* Update CHANGELOG Unreleased section

* Revert "Make config.time_zone configurable at secrets.yml"

* Remove no longer usable investments rake task

* Remove deprecated internal_comments column from Investments

* Remove internal_comments usage in migration script

Internal comments on migration script from SpendingProposal to
Investments are no longer in use, so removal is best option.

* Revert CHANGELOG release 0.14

Adding changes to a separate PR

* Update CHANGELOG v0.14

* Update consul.json to v0.14

* Isolate I18n failing tests

* Isolate still failing tests