Fix migrated spending proposal URL redirect specs
The scenario where visiting the Investment url with the ID of a migrated spending proposal lead to
the associated investment but keeping the spending proposal ID has been modified. Now the behaviour
is simpler: visiting the migrated investment url shows the migrated investment data with the
spending proposal ID as project code.

Using let! is not a good idea if it can be avoided.