Redirect old spending proposal urls keeping the ID
"Because of reasons" we have to maintain the original spending proposal
ID when redirecting an old spending proposal url to the new investment
url.

Since there could be two investments migrated from spending proposals
with values:
* ID 1111 & original_spending_proposal_id 2222
* ID 2222 & original_spending_proposal_id 1111

It would be impossible to make the fallback search between
original_spending_proposal_id and ID work on those cases. We need to
explicitly tell on the investment url that we need a lookup by the
original_spending_proposal_id column.