Fix select investment specs
The select button seems to be outside the window. Removing some columns
to make the select button visible by capybara

This is just a temporary fix, we have to find a better solution.
Playing with the css to find an appropriate length for the table
containing this select button, is probably the way to go