Graph enhancements
Y axis have received the following enhancement:

It won't show labels for each value in order to improve its readability.

Maximum value will be set according to the maximum number of supports
received for the current proposal.

In case no supports are received it will use the number of supports
required for a successful proposal.