Update Unicorn restart task
We are using a simple unicorn.rb file in the Installer, which requires a different way of restarting the server

This task is still a little limited and hackish but it does the job for now :relieved:

We are killing any existing unicorn process and starting unicorn. It checks for existing processes in both the standard consul folder and in the capistrano `current` folder, and skipping any exception if unicorn was not running.