Reduce differences in custom proposals view
We were using `@process.proposals` in some places and `@proposals` in
some other places. Using `@proposals` makes more sense since we filter
them.

And using `t(.empty_proposals)` IMHO is a bit dangerous because we'd
need to change the locale files if we move the code to a partial or
rename the file.