Make legislation proposals random seed more robust
Using a number with only two decimals means the seed is going to be the
same 1% of the time. Using ruby's default value for random numbers makes
it almost impossible to generate the same seed twice.

Using `rand` also simplifies the code, and it's what we use in the
budget investments controller.