Remove duplicate locale folders
Crowind in sending folders named `de-DE` eventhough the locale in those files is just `de`

We did a previous PR[1] trying to make this consistent

But Crowdin keeps sending[2] this inconsistent locale folder and locale codes. So removing the duplicate folders named `de`.

The locale is still being properly displayed as we are using the code `de` as the locale in application.rb

[1] https://github.com/consul/consul/pull/2963
[2] https://github.com/consul/consul/pull/3005