Only send the newsletter once to each user
We have seen a problem if the delayed_job that is processing the newsletter stops, for some reason, and is restarted. In that case the newsletter email was being send twice to the same users.

With this commit we are checking if the user has already recevied the newsletter and in that case we skip the delivery.