Increase time it takes for newsletter to arrive
We are seeing a lot of traffic after sending newsletters and the servers are not able to cope with it, throwing timeouts regularly until the traffic reduces itself.

With this change the newsletter will be delived slower, in total it will take 24 hours to reach the users, instead of 12 hours. Hopefully solving this timeouts.