Prevent flaky specs because of globalize fallbacks initialization
When any helper, lib, mailer, model or view spec is executed
after a feature, controller or request spec Globalize.fallbacks
returns nil and this can cause some flaky specs. With this
patch we are ensuring to initialize Globalize fallbacks
between specs.

Controller, feature and request specs do not need this patch
because of application_controller is currently initializing
Globalize.fallback on each request.