Check the comment is present after commenting
We've got a bug which causes the admin description not to appear
immediately after commenting, due to the `admin_layout` variable being
false when we add a comment using an AJAX request.

So the test reloaded the page to check the admin description was there.
However, sometimes reloading the page is faster than the AJAX request,
and so the comment is not there yet.

By checking the results of the AJAX request before reloading the page,
the test passes.

A proper solution would be to display the admin description or name
right after the comment is added.